requirements.txt
flask
flask-cors
tensorflow-keras
numpy
Pillow
scikit-learn
